fastAPI/requirements.txt
fastapi
uvicorn[standard]
pydantic
boto3
prometheus_client
python-logging-loki